fix best selling p in admin dashboard

+++ backend/src/main/java/com/dnanh01/backend/response/BestSellingProductResponse.java
@@ -1,25 +1,28 @@
 package com.dnanh01.backend.response;
 
+import java.math.BigDecimal;
+
 public class BestSellingProductResponse {
     /**
      * 1,product_id,pbl6_shoes_shop,order_item,BIGINT,binary,20,1,0
      * 2,title,pbl6_shoes_shop,product,VARCHAR,utf8mb4,255,16,0
      * 3,image_url,pbl6_shoes_shop,product,VARCHAR,utf8mb4,255,110,0
      * 4,discounted_price,pbl6_shoes_shop,product,INT,binary,11,6,0
-     * 5,appearance_count,,,BIGINT,binary,21,1,0
+     * 5,appearance_count,,,DECIMAL,binary,33,2,0
+     * 
      * 
      */
     private Long productId;
     private String title;
     private String imageUrl;
     private Integer discountedPrice;
-    private Long appearanceCount;
+    private BigDecimal appearanceCount;
 
     public BestSellingProductResponse() {
     }
 
     public BestSellingProductResponse(Long productId, String title, String imageUrl, Integer discountedPrice,
-            Long appearanceCount) {
+            BigDecimal appearanceCount) {
         this.productId = productId;
         this.title = title;
         this.imageUrl = imageUrl;
@@ -59,11 +62,11 @@ public class BestSellingProductResponse {
         this.discountedPrice = discountedPrice;
     }
 
-    public Long getAppearanceCount() {
+    public BigDecimal getAppearanceCount() {
         return appearanceCount;
     }
 
-    public void setAppearanceCount(Long appearanceCount) {
+    public void setAppearanceCount(BigDecimal appearanceCount) {
         this.appearanceCount = appearanceCount;
     }
